requirements.txt
Django==3.2
django-cleanup==5.2.0
environs==9.3.5
folium==0.12.1.post1
Pillow==8.1.2
random-username==1.0.2
phonenumbers==8.12.38
django-phonenumber-field==6.0.0
python-dateutil==2.8.2
